+++ docs/source/changelog.rst
@@ -1,8 +1,24 @@
 Changelog
 ===========
 
+.. raw:: html
+   :file: ./_includes/official_links.html
+
 All notable changes to this project are documented here. The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
+
+v0.5.0 (2024-05-04)
+---------------------
+
+Added
+^^^^^^^^^
+- Recursive chunking strategy.
+- Documentation updated.
+
+Changed
+^^^^^^^^^
+- Saving chunks to disk option has been removed, as this is something that can be done outside of the library.
+
 
 v0.4.0 (2024-04-17)
 ---------------------